spike(PortDemo): Added icons
